+++ packages/nx-dx-docker-plugin/src/executors/docker-push/schema.ts
@@ -1,0 +1,10 @@
+import {
+  type DockerRunOptions,
+  dockerRunOptionsSchema,
+} from "../../docker-run.ts";
+
+export const dockerPushExecutorSchema = dockerRunOptionsSchema;
+
+export type DockerPushExecutorInput = Partial<DockerPushExecutorSchema>;
+
+export type DockerPushExecutorSchema = DockerRunOptions;